chore: refactor list_collections
